fix: try to serve files with request path before redirects
Before this commit, all requests were first checked for a redirect.
This involved checking the current redirect map, then updating
it where there was no match.

This meant that *every time* a file was requested, the redirect
map was pulled - harming the performance of the server.

This change ensures that gosherve first tries to match a file
path and serve a file, before then checking for a redirect
match.