perf: use http.ServeFileFS to serve files without fs.ReadFile
Previously, files were served with http.ServeContent, which took
an io.ReadSeeker in the form of a bytes.Buffer. This meant reading
every requested file into memory, causing many, many more allocations
on the heap than were required, and slowing the request throughput
on the server.

This change uses the (newly introduced) http.ServeFileFS to serve
the file, which in turn uses io.Copy on the underlying files in the
filesystem - which allocates a single buffer which fits on the stack.