fix: improve ETag generation
In the previous iteration, if a request was made to a directory,
and thus the index.html was served, the size component of the ETag
was reported as `0` since directories have a zero size.

This breaks the ETag generation out into a function which takes
a filename/filesystem and generates an appropriate ETag.